Build with core_current
This lib has been built with core-libart and core-oj which include
private APIs. In order to prevent this lib from depending on the private
APIs in the future, directly dependency to the core libraries are
replaced with the pseudo-sdk stub called 'core_current' that has public
core Java APIs only.

Bug: 72206056
Test: mma -j under this directory
Change-Id: I7f15c3b4aca1bdf1b57a5457322f7a01ddccee88